Add function to check form data in request
A new function has been added to verify if the content type in a request needs form data. The function, 'needsFormData', checks if the content type is either 'application/x-www-form-urlencoded' or 'multipart/form-data'. Also, changes have been made in the parsing of request body such that, only if the content type needs form data and the method is either POST, PUT, or PATCH, the body bytes are read.